Amend Rush model and migration
Changes a number of fields from integers to floats.

+++ lib/nfl_rushing/stats/rush.ex
@@ -4,9 +4,9 @@ defmodule NflRushing.Stats.Rush do
 
   schema "rushes" do
     field :attempts, :integer
-    field :attempts_per_game, :integer
-    field :avg_yards_per_attempt, :integer
-    field :first_down_percentage, :integer
+    field :attempts_per_game, :float
+    field :avg_yards_per_attempt, :float
+    field :first_down_percentage, :float
     field :first_downs, :integer
     field :forty_plus, :integer
     field :fumbles, :integer
@@ -18,7 +18,7 @@ defmodule NflRushing.Stats.Rush do
     field :total_yards, :integer
     field :touchdowns, :integer
     field :twenty_plus, :integer
-    field :yards_per_game, :integer
+    field :yards_per_game, :float
 
     timestamps()
   end
@@ -26,7 +26,41 @@ defmodule NflRushing.Stats.Rush do
   @doc false
   def changeset(rush, attrs) do
     rush
-    |> cast(attrs, [:player_name, :team_abbr, :position, :attempts_per_game, :attempts, :total_yards, :avg_yards_per_attempt, :yards_per_game, :touchdowns, :longest, :longest_is_touchdown, :first_downs, :first_down_percentage, :twenty_plus, :forty_plus, :fumbles])
-    |> validate_required([:player_name, :team_abbr, :position, :attempts_per_game, :attempts, :total_yards, :avg_yards_per_attempt, :yards_per_game, :touchdowns, :longest, :longest_is_touchdown, :first_downs, :first_down_percentage, :twenty_plus, :forty_plus, :fumbles])
+    |> cast(attrs, [
+      :player_name,
+      :team_abbr,
+      :position,
+      :attempts_per_game,
+      :attempts,
+      :total_yards,
+      :avg_yards_per_attempt,
+      :yards_per_game,
+      :touchdowns,
+      :longest,
+      :longest_is_touchdown,
+      :first_downs,
+      :first_down_percentage,
+      :twenty_plus,
+      :forty_plus,
+      :fumbles
+    ])
+    |> validate_required([
+      :player_name,
+      :team_abbr,
+      :position,
+      :attempts_per_game,
+      :attempts,
+      :total_yards,
+      :avg_yards_per_attempt,
+      :yards_per_game,
+      :touchdowns,
+      :longest,
+      :longest_is_touchdown,
+      :first_downs,
+      :first_down_percentage,
+      :twenty_plus,
+      :forty_plus,
+      :fumbles
+    ])
   end
 end
