+++ priv/repo/migrations/20200830231721_create_rushes.exs
@@ -6,16 +6,16 @@ defmodule NflRushing.Repo.Migrations.CreateRushes do
       add :player_name, :string
       add :team_abbr, :string
       add :position, :string
-      add :attempts_per_game, :integer
+      add :attempts_per_game, :float
       add :attempts, :integer
       add :total_yards, :integer
-      add :avg_yards_per_attempt, :integer
-      add :yards_per_game, :integer
+      add :avg_yards_per_attempt, :float
+      add :yards_per_game, :float
       add :touchdowns, :integer
       add :longest, :integer
       add :longest_is_touchdown, :boolean, default: false, null: false
       add :first_downs, :integer
-      add :first_down_percentage, :integer
+      add :first_down_percentage, :float
       add :twenty_plus, :integer
       add :forty_plus, :integer
       add :fumbles, :integer
